requirements_old.txt
python-telegram-bot==13.15
openai==0.28.1
python-dotenv==1.0.0
requests==2.31.0
